+++ react/src/lib/components/SmartNodeSelector/components/SmartNodeSelectorComponent.tsx
@@ -123,8 +123,7 @@ export default class SmartNodeSelectorComponent extends Component<SmartNodeSelec
         this.mouseDownElement = null;
         this.componentIsMounted = false;
 
-        let hasError = false;
-        let error = "";
+        let error: string | undefined;
         try {
             this.treeData = new TreeData({
                 treeData: props.data,
@@ -132,7 +131,6 @@ export default class SmartNodeSelectorComponent extends Component<SmartNodeSelec
             });
         } catch (e) {
             this.treeData = null;
-            hasError = true;
             error = e;
         }
 
@@ -154,11 +152,11 @@ export default class SmartNodeSelectorComponent extends Component<SmartNodeSelec
             nodeSelections,
             currentTagIndex: 0,
             suggestionsVisible: false,
-            hasError: hasError,
-            error: error,
+            hasError: error !== undefined,
+            error: error || "",
         };
 
-        if (!hasError) {
+        if (error === undefined) {
             this.numValidSelections = this.countValidSelections();
         } else {
             this.numValidSelections = 0;
@@ -215,6 +213,43 @@ export default class SmartNodeSelectorComponent extends Component<SmartNodeSelec
     }
 
     componentDidUpdate(prevProps: SmartNodeSelectorPropsType): void {
+        if (
+            (this.props.data &&
+                JSON.stringify(this.props.data) !==
+                    JSON.stringify(prevProps.data)) ||
+            (this.props.delimiter &&
+                this.props.delimiter !== prevProps.delimiter)
+        ) {
+            let error: string | undefined;
+            try {
+                this.treeData = new TreeData({
+                    treeData: this.props.data,
+                    delimiter: this.props.delimiter,
+                });
+            } catch (e) {
+                this.treeData = null;
+                error = e;
+            }
+            const nodeSelections: TreeNodeSelection[] = [];
+            for (const node of this.state.nodeSelections) {
+                nodeSelections.push(
+                    this.createNewNodeSelection(node.getNodePath())
+                );
+            }
+
+            this.setState(
+                {
+                    nodeSelections: nodeSelections,
+                    currentTagIndex: this.state.currentTagIndex,
+                    suggestionsVisible: this.state.suggestionsVisible,
+                    hasError: error !== undefined,
+                    error: error || "",
+                },
+                () => {
+                    this.updateSelectedTagsAndNodes();
+                }
+            );
+        }
         const selectedTags = this.state.nodeSelections
             .filter((nodeSelection) => nodeSelection.isValid())
             .map((nodeSelection) =>